cap nhat toan nang

+++ DoAnPTUD/DTO/DTO_ChiTietGiaoDich.cs
@@ -9,13 +9,13 @@ namespace DTO
     public class DTO_ChiTietGiaoDich
     {
         private int MaGD;
-        private int SoTKNguoiChuyen;
-        private int SoTKNguoiNhan;
+        private long SoTKNguoiChuyen;
+        private long SoTKNguoiNhan;
         private float SoTien;
         private DateTime NgayGio;
         private string DienGia;
 
-        public DTO_ChiTietGiaoDich(int maGD, int soTKNguoiChuyen, int soTKNguoiNhan, float soTien, DateTime ngayGio, string dienGia)
+        public DTO_ChiTietGiaoDich(int maGD, long  soTKNguoiChuyen, long soTKNguoiNhan, float soTien, DateTime ngayGio, string dienGia)
         {
             MaGD = maGD;
             SoTKNguoiChuyen = soTKNguoiChuyen;
@@ -24,7 +24,7 @@ namespace DTO
             NgayGio = ngayGio;
             DienGia = dienGia;
         }
-        public DTO_ChiTietGiaoDich( int soTKNguoiChuyen, int soTKNguoiNhan, float soTien, DateTime ngayGio, string dienGia)
+        public DTO_ChiTietGiaoDich( long soTKNguoiChuyen, long soTKNguoiNhan, float soTien, DateTime ngayGio, string dienGia)
         {
             SoTKNguoiChuyen = soTKNguoiChuyen;
             SoTKNguoiNhan = soTKNguoiNhan;
@@ -36,8 +36,8 @@ namespace DTO
         }    
 
         public int MaGD1 { get => MaGD; set => MaGD = value; }
-        public int SoTKNguoiChuyen1 { get => SoTKNguoiChuyen; set => SoTKNguoiChuyen = value; }
-        public int SoTKNguoiNhan1 { get => SoTKNguoiNhan; set => SoTKNguoiNhan = value; }
+        public long SoTKNguoiChuyen1 { get => SoTKNguoiChuyen; set => SoTKNguoiChuyen = value; }
+        public long SoTKNguoiNhan1 { get => SoTKNguoiNhan; set => SoTKNguoiNhan = value; }
         public float SoTien1 { get => SoTien; set => SoTien = value; }
         public DateTime NgayGio1 { get => NgayGio; set => NgayGio = value; }
         public string DienGia1 { get => DienGia; set => DienGia = value; }
